Reverted #234 DI\object() extend the previous definition again
