+++ src/DI/Definition/Helper/ClassDefinitionHelper.php
@@ -12,7 +12,6 @@ namespace DI\Definition\Helper;
 use DI\Definition\ClassDefinition;
 use DI\Definition\ClassDefinition\MethodInjection;
 use DI\Definition\ClassDefinition\PropertyInjection;
-use DI\Definition\ObjectDefinitionExtension;
 use DI\Scope;
 
 /**
@@ -56,23 +55,14 @@ class ClassDefinitionHelper implements DefinitionHelper
     private $methods = array();
 
     /**
-     * @var bool
-     */
-    private $extension;
-
-    /**
      * Helper for defining an object.
      *
      * @param string|null $className Class name of the object.
      *                               If null, the name of the entry (in the container) will be used as class name.
-     * @param bool        $extension Does this definition extend another?
-     *
-     * @return ClassDefinitionHelper
      */
-    public function __construct($className = null, $extension = false)
+    public function __construct($className = null)
     {
         $this->className = $className;
-        $this->extension = $extension;
     }
 
     /**
@@ -203,11 +193,7 @@ class ClassDefinitionHelper implements DefinitionHelper
      */
     public function getDefinition($entryName)
     {
-        if ($this->extension) {
-            $definition = new ObjectDefinitionExtension($entryName, $this->className);
-        } else {
-            $definition = new ClassDefinition($entryName, $this->className);
-        }
+        $definition = new ClassDefinition($entryName, $this->className);
 
         if ($this->lazy !== null) {
             $definition->setLazy($this->lazy);
